+++ packages/gatsby-plugin-feed/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [3.0.0-next.0](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-feed@2.14.0-next.0...gatsby-plugin-feed@3.0.0-next.0) (2021-02-05)
+
+**Note:** Version bump only for package gatsby-plugin-feed
+
+# [2.14.0-next.0](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-feed@2.13.0-next.0...gatsby-plugin-feed@2.14.0-next.0) (2021-01-28)
+
+**Note:** Version bump only for package gatsby-plugin-feed
+
 # [2.13.0-next.0](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-feed@2.12.0-next.1...gatsby-plugin-feed@2.13.0-next.0) (2021-01-18)
 
 **Note:** Version bump only for package gatsby-plugin-feed